requirements.txt
alembic==1.5.7
cffi==1.14.5
click==7.1.2
cryptography==3.4.6
ecdsa==0.14.1
fastapi==0.63.0
greenlet==1.0.0
h11==0.12.0
Mako==1.1.4
MarkupSafe==1.1.1
passlib==1.7.4
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-jose==3.2.0
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.4.1
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.13.4
